+++ client/src/index.tsx
@@ -4,8 +4,8 @@ initSentry();
 /* @refresh reload */
 import { render } from "solid-js/web";
 import { Router } from "@solidjs/router";
-import "virtual:uno.css";
 import "@unocss/reset/tailwind.css";
+import "virtual:uno.css";
 import "./styles/global.css";
 import "./styles/themes-structure.css";
 import "./styles/themes.css";